+++ app/client/js/editor.js
